# static/openapi.yaml
openapi: 3.1.0
info:
  title: EmptyInbox API
  description: |
    Disposable email inbox API for AI agents and developers.

    Create temporary email addresses, receive messages, and read content — all via REST.
    No browser required. Authenticate with a static API key.

    **Zero-config agent workflow:**
    1. `POST /auth/register` — create a free account and get an API key (no passkey needed)
    2. `POST /inbox` — create a disposable address
    3. Trigger an external signup/verification using that address
    4. `GET /messages` — poll until the email arrives
    5. Read `code` or `action_url` off the message — the API extracts both for you

    Or use the [emptyinbox-mcp](https://www.npmjs.com/package/emptyinbox-mcp) package — auto-registers on first run, no setup needed.

    Human accounts: get your API key at https://emptyinbox.me/settings.html
  version: 1.2.0
  contact:
    url: https://emptyinbox.me
servers:
  - url: https://emptyinbox.me/api
    description: Production
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        Your API key. Obtain one by calling `POST /auth/register` (agents) or from https://emptyinbox.me/settings.html (humans).
        Usage: `Authorization: Bearer <your_api_key>`
  schemas:
    PaymentQuote:
      type: object
      properties:
        address:
          type: string
          description: Bitcoin address to pay. Unique to this quote.
          example: bc1qexampleaddress
        bundle:
          type: string
          example: starter
        quota:
          type: integer
          description: Inboxes granted once payment is seen
          example: 50
        usd:
          type: integer
          example: 5
        amount_satoshis:
          type: integer
          example: 5300
        amount_btc:
          type: string
          example: "0.00005300"
        bip21:
          type: string
          description: Wallet URI with address and amount prefilled
          example: "bitcoin:bc1qexampleaddress?amount=0.00005300"
        expires_at:
          type: string
          example: "2026-09-09T12:15:00Z"
        status_url:
          type: string
          example: /api/payments/status/bc1qexampleaddress
...
paths:
...
  /inbox:
    post:
      operationId: createInbox
      summary: Create a new disposable email inbox
      description: |
        Creates a new inbox and returns its email address as plain text.
        Each inbox consumes 1 from your quota. All accounts start with 5 inboxes.
        When quota runs out this returns 402 with a link to POST /payments/quote,
        which sells more quota for Bitcoin. Humans can also buy at https://emptyinbox.me/inboxes.html
      responses:
        "201":
          description: Inbox created — returns the email address
          content:
            text/plain:
              schema:
                type: string
              example: clever.sunny.butterfly@emptyinbox.me
        "401":
          description: Missing or invalid API key
        "402":
          description: |
            Inbox quota exhausted. The body carries the URLs for buying more.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: insufficient_quota
                  message:
                    type: string
                  bundles_url:
                    type: string
                    example: /api/payments/bundles
                  quote_url:
                    type: string
                    example: /api/payments/quote
...
  /payments/quote:
    post:
      operationId: createQuote
      summary: Quote a Bitcoin price for a quota bundle
      description: |
        Returns a payment address, the exact amount to send, and a BIP21 URI.
        Send the amount from any Bitcoin wallet, then poll
        GET /payments/status/{address}.

        Quota is credited as soon as the transaction is seen on the network,
        without waiting for confirmations, unless the transaction signals
        Replace-By-Fee. That credit is provisional: if the transaction has not
        confirmed within 6 hours it is reversed.

        Idempotent per account. Calling it again while a quote is open returns
        the same address rather than issuing a new one.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  description: Bundle id from /payments/bundles
                  example: starter
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentQuote"
        "200":
          description: An open quote already existed and was returned unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentQuote"
        "400":
          description: Unknown bundle id
        "401":
          description: Missing or invalid API key
        "409":
          description: A previous payment is still awaiting confirmation
        "502":
          description: Payment provider unreachable
...